add CU info to dynamic-regions c++ API
