+++ src/runtime_src/core/common/info_memory.h
@@ -28,7 +28,7 @@ boost::property_tree::ptree
 memory_topology(const xrt_core::device * device);
 
 boost::property_tree::ptree
-xclbin_info(const xrt_core::device * device);
+dynamic_regions(const xrt_core::device * device);
 
 }} // memory, xrt
